requirements.txt
fastapi[standard]
uvicorn[standard]
transformers
pillow
torch
python-multipart